celery_worker/requirements.txt
runpod
requests
sqlalchemy
psycopg2-binary
python-dotenv
boto3
celery
redis
requests
